+++ AlchemistCode/Cards/Rare/Refine.cs
@@ -3,6 +3,7 @@ using Alchemist.AlchemistCode.Powers;
 using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.HoverTips;
 
 namespace Alchemist.AlchemistCode.Cards.Rare;
 
@@ -13,7 +14,8 @@ public class Refine : AlchemistCard
     {
         WithCostUpgradeBy(-1);
         WithQuietPower<RefinePower>(1, 0);
-        WithTips(_ => Mixing.MixTips(upgraded: true));
+        WithTips(_ => Mixing.MixRefTips());
+        WithTips(_ => new[] { HoverTipFactory.FromKeyword(CardKeyword.Retain) });
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)